pet_monitor/requirements.txt
Flask>=2.3.0
numpy>=1.24.0
pyalsaaudio>=0.10.0
scipy>=1.11.0
watchdog>=3.0.0
